move print matrix to utils
